+++ tests/e2e/fixtures/d1_crud/cloesce.config.ts
@@ -1,6 +1,0 @@
-import { CloesceConfigOptions } from "cloesce";
-export default {
-    srcPaths: ["./"],
-    workersUrl: "http://localhost:5391/api",
-} satisfies CloesceConfigOptions;
-