+++ tests/e2e/fixtures/data_source_type/cloesce.config.ts
@@ -1,6 +1,0 @@
-import { CloesceConfigOptions } from "cloesce";
-export default {
-    srcPaths: ["./"],
-    workersUrl: "http://localhost:5719/api",
-} satisfies CloesceConfigOptions;
-