+++ tests/e2e/tests/middleware.test.ts
@@ -1,14 +1,14 @@
 import { startWrangler } from "../src/setup.js";
 import { describe, it, expect, beforeAll, afterAll } from "vitest";
 import { Foo } from "../fixtures/middleware/client";
-import config from "../fixtures/middleware/cloesce.config";
+import config from "../fixtures/middleware/cloesce.jsonc" with { type: "jsonc" };
 
 let stopWrangler: () => Promise<void>;
 beforeAll(async () => {
   // NOTE: e2e is called from proj root
   stopWrangler = await startWrangler(
     "./fixtures/middleware",
-    config.workersUrl!,
+    config.workers_url!,
   );
 }, 30_000);