+++ tests/e2e/tests/poos.test.ts
@@ -1,12 +1,12 @@
 import { startWrangler, withRes } from "../src/setup.js";
 import { describe, it, expect, beforeAll, afterAll } from "vitest";
 import { PooAcceptYield } from "../fixtures/poos/client";
-import config from "../fixtures/poos/cloesce.config";
+import config from "../fixtures/poos/cloesce.jsonc" with { type: "jsonc" };
 
 let stopWrangler: () => Promise<void>;
 beforeAll(async () => {
   // NOTE: e2e is called from proj root
-  stopWrangler = await startWrangler("./fixtures/poos", config.workersUrl!);
+  stopWrangler = await startWrangler("./fixtures/poos", config.workers_url!);
 }, 30_000);
 
 afterAll(async () => {